requirements.txt
Flask==3.0.0
openai==1.61.1
#langchain-openai==0.3.4
#gunicorn==23.0.0
azure-identity
#azure-identity==1.19.0
azure-core
azure-keyvault-secrets
azure-search-documents
#
